+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Rename `Tree::node_by_svg_id` to `Tree::node_by_id`.
+- Use `rctree::Node<NodeKind>` instead of `rctree::Node<Box<NodeKind>>`.
+
+### Removed
+- `NodeExt::kind`. Use `Node::borrow` instead.
 
 ### Fixed
 - Panic during `visibility` resolving.